### 2019-11-13 #1888 @labarba
👋 @danielskatz — This submission falls in the general area of software engineering, and I think you could handle it as editor?
*Hypothesis: A new approach to property-based testing*
https://github.com/HypothesisWorks/hypothesis/
