### 2019-11-13 #1888 @luizirber
I was going to volunteer as a reviewer, but I skimmed the paper and it cites hypothesis-bio as an example (and I'm involved in the development). Is it a conflict of interest?
 https://github.com/HypothesisWorks/hypothesis/blob/master/paper.md
