### 2019-11-13 f0fd22e
Excluding zero does both signs
