### 2019-11-13 #1888 @danielskatz
👋 @DRMacIver - Can you suggest some additional reviewers? From [this list](https://bit.ly/joss-reviewers) perhaps?
