### 2019-11-13 #1888 @DRMacIver
Sure!

Based solely on the process of looking for "testing" in the topics and "Python" in the primary language and excluding @Zac-HD because he's one of the authors, the following seem to be the valid choices of reviewers from that spreadsheet: @ctb @djmitche @camillescott

I don't believe I know any of them personally, and based on looking at their github profiles it doesn't look like any of them currently use Hypothesis, so I don't have any criteria I'd use to narrow it down further than that and I guess any of those three who are interested would be good choices.
