### 2019-11-13 #1888 @danielskatz
👋 @djmitche - would you be willing to review this for JOSS?
