### 2019-11-13 #1888 @Zac-HD
👋 everyone!  Super excited to go from JOSS admirer to JOSS author 😁

The issue tags seem a little weird - we do have all those languages in the repo, but Hypothesis is a pure-Python library.  For later stages, we've been archiving Hypothesis at [10.5281/zenodo.1412597](https://doi.org/10.5281/zenodo.1412597)
