### 2019-11-13 #1888 @danielskatz
The archive isn't needed until later, and there may be changes that needed to be made and put in a new release/archive...
