### 2019-11-14 #1888 @Zac-HD
No worries, [we automated the release process](https://hypothesis.works/articles/continuous-releases/) years ago, so that's as easy as  "approve and merge a PR".
