### 2019-11-14 #1888 @djmitche
> wave @djmitche - would you be willing to review this for JOSS?

I would.  I'm a very casual user of hypothesis, so I'm excited to see it in JOSS.
