### 2019-11-14 #1891 @whedon
[REVIEW]: Hypothesis: A new approach to property-based testing
**Submitting author:** @DRMacIver (<a href="http://orcid.org/0000-0002-8635-3223">David MacIver</a>)
**Repository:** <a href="https://github.com/HypothesisWorks/hypothesis/" target ="_blank">https://github.com/HypothesisWorks/hypothesis/</a>
**Version:** 4.45.0
