### 2019-11-14 c3f7395
Warning-free async tests
