+++ hypothesis-python/tests/py3/test_asyncio.py
@@ -19,7 +19,6 @@ from __future__ import absolute_import, division, print_function
 
 import asyncio
 import sys
-import unittest
 from unittest import TestCase
 
 import pytest
@@ -27,6 +26,11 @@ import pytest
 import hypothesis.strategies as st
 from hypothesis import assume, given
 from hypothesis.internal.compat import PYPY
+
+if sys.version_info < (3, 8):
+    coro_decorator = asyncio.coroutine
+else:
+    coro_decorator = pytest.mark.skip
 
 
 class TestAsyncio(TestCase):
@@ -59,9 +63,8 @@ class TestAsyncio(TestCase):
             raise error
 
     @pytest.mark.skipif(PYPY, reason="Error in asyncio.new_event_loop()")
-    @pytest.mark.skipif(sys.version_info[:2] >= (3, 8), reason="deprecated @coroutine")
     @given(st.text())
-    @asyncio.coroutine
+    @coro_decorator
     def test_foo(self, x):
         assume(x)
         yield from asyncio.sleep(0.001)
@@ -77,12 +80,8 @@ class TestAsyncioRun(TestCase):
 
     @pytest.mark.skipif(sys.version_info[:2] < (3, 7), reason="asyncio.run() is new")
     @given(st.text())
-    @asyncio.coroutine
+    @coro_decorator
     def test_foo(self, x):
         assume(x)
         yield from asyncio.sleep(0.001)
         assert x
-
-
-if __name__ == "__main__":
-    unittest.main()
