### 2019-11-14 e27da9b
Infer for Final, TypedDict types
