+++ hypothesis-python/tests/conftest.py
@@ -33,6 +33,8 @@ run()
 # Skip collection of tests which require the Django test runner,
 # or that don't work on the current major version of Python.
 collect_ignore_glob = ["django/*", "py3/*" if sys.version_info[0] == 2 else "py2/*"]
+if sys.version_info < (3, 8):
+    collect_ignore_glob.append("py3/*py38*")
 
 
 def pytest_configure(config):
