### 2019-11-16 8a7701b
Add and annotate no-branch pragma
