### 2019-11-17 #1891 @djmitche
@danielskatz I think I see an error in the paper itself ("exception" instead of "extension"), but I'm not sure where to find the source for that paper.  Pointers?
