### 2019-11-17 #1891 @danielskatz
see paper.md and paper.bib in the source repo (root level)
