### 2019-11-17 HypothesisWorks/hypothesis#2210 @djmitche
Minor issues with JOSS paper
> Additionally, Hypothesis is easily extensible, and has a number of user-provided exceptions for specific research applications.

I think that "exceptions" should be "extensions"?

> Given some interestingness test (e.g. that it triggers a bug in some software), how do we generate a "good" test case that passes that interestingness test?

Use of the term "test" for two different purposes in this sentence is confusing.  Consider "interestingness condition"?

(refs https://github.com/openjournals/joss-reviews/issues/1891)
