### 2019-11-18 3fe2eaa
Use TypedDict patch
