### 2019-11-18 56c25e9
De-flake timing test
