### 2019-11-18 #1891 @luizirber
Only one nitpick: There is nothing wrong with the LICENSE.txt file per se, but if you put the content of the MPL 2.0 license linked it will show up as `MPL 2.0` in the repository landing page (instead of `View license`, like it is now).

All my checkboxes are checked.

And I must say this was the easiest review I ever did, but it was somewhat expected because I tend to look for hypothesis dev practices as reference for my projects. The paper is also well written and does a great intro to property testing for scientific projects. Kudos @DRMacIver and @Zac-HD!
