### 2019-11-18 #1891 @Zac-HD
Thanks @luizirber :tada:

For the licence, I think this is mostly an effect of Github not having fantastic support for monorepos... [`hypothesis-python/LICENSE.txt`](https://github.com/HypothesisWorks/hypothesis/blob/master/hypothesis-python/LICENSE.txt) is the file that gets packaged up, and it's the standard MPL2.0 text.  We also use the `setuptools` license classifier, and have a copyright+license header in every `.py` file :smile:
