### 2019-11-19 #1891 @danielskatz
Thanks all - I'll continue the process, first checking the license issue, then proof-reading and addressing the final steps
