### 2019-11-19 #1891 @danielskatz
👋 @arfon - Can you take a look at the license discussion here and the files in the repo and confirm that this is ok, or if not, suggest what changes need to be made?
