### 2019-11-19 #1891 @arfon
> 👋 @arfon - Also, is there any issue with the author list?

It's unconventional, but I think it's probably OK.

> 👋 @arfon - Can you take a look at the license discussion here and the files in the repo and confirm that this is ok, or if not, suggest what changes need to be made?

I also think this is OK.
