### 2019-11-19 9b11f4e
minor changes for JOSS paper
