+++ paper.md
@@ -31,7 +31,7 @@ affiliations:
 *Property-based testing* is a style of testing popularised by the QuickCheck family of libraries,
 first in Haskell [@DBLP:conf/icfp/ClaessenH00] and later in Erlang [@DBLP:conf/erlang/ArtsHJW06],
 which integrates generated test cases into existing software testing workflows:
-Instead of tests which provide examples of a single concrete behaviour,
+Instead of tests that provide examples of a single concrete behaviour,
 tests specify properties that hold for a wide range of inputs,
 which the testing library then attempts to generate test cases to refute.
 For a general introduction to property-based testing, see [@PraiseOfPBT].
@@ -63,8 +63,8 @@ This is true both because of the much wider array of software that can now easil
 and because it has a novel implementation that lifts a major difficulty that prior research on software testing faces.
 
 Much of software testing research boils down to variants on the following problem:
-Given some interestingness condition (e.g. that it triggers a bug in some software),
-how do we generate a "good" test case which satisfies that condition?
+Given some interestingness condition (e.g., that it triggers a bug in some software),
+how do we generate a "good" test case that satisfies that condition?
 
 Particular sub-problems of this are:
 
@@ -94,7 +94,7 @@ and improvements to the generation process can operate solely on this universal 
 Currently Hypothesis uses this format to support two major use cases:
 
 1. It is the basis of its approach to test-case reduction, allowing it to support more powerful test-case reduction than is found in most property-based testing libraries with no user intervention.
-2. It supports Targeted Property-Based Testing [@DBLP:conf/issta/LoscherS17], which uses a score to guide testing towards a particular goal (e.g. maximising an error term). In the original implementation this would require custom mutation operators per test,
+2. It supports Targeted Property-Based Testing [@DBLP:conf/issta/LoscherS17], which uses a score to guide testing towards a particular goal (e.g., maximising an error term). In the original implementation this would require custom mutation operators per test,
    but in Hypothesis this mutation is transparent to the user and they need only specify the goal.
 
 The internal format is flexible and contains rich information about the structure of generated test cases,
