### 2019-11-19 #1891 @danielskatz
@DRMacIver & @Zac-HD - regarding the wording in the paper:

- "lifts a major difficulty" is an odd phrase to me - could we uses resolves or addresses instead of lifts?

Also, please see https://github.com/HypothesisWorks/hypothesis/pull/2212
