### 2019-11-19 f6aff4c
Add function that uses check_type() to validate arguments used by Rule, this will safely raise an InvalidArgument error instead of an internal error.
