+++ hypothesis-python/src/hypothesis/stateful.py
@@ -364,6 +364,12 @@ def multiple(*args):
     return MultipleResults(args)
 
 
+def validate_rule_arguments(kwargs):
+    """Validates arguments used by Rule"""
+    for k, v in kwargs.items():
+        check_type(SearchStrategy, v, k)
+
+
 def _convert_targets(targets, target):
     """Single validator and convertor for target arguments."""
     if target is not None:
@@ -432,6 +438,7 @@ def rule(targets=(), target=None, **kwargs):
     provided.
     """
     converted_targets = _convert_targets(targets, target)
+    validate_rule_arguments(kwargs)
 
     def accept(f):
         existing_rule = getattr(f, RULE_MARKER, None)
@@ -467,6 +474,7 @@ def initialize(targets=(), target=None, **kwargs):
     arbitrary order.
     """
     converted_targets = _convert_targets(targets, target)
+    validate_rule_arguments(kwargs)
 
     def accept(f):
         existing_rule = getattr(f, RULE_MARKER, None)
