### 2019-11-19 afe3d6d
Add testing for valid and invalid rule arguments.
