### 2019-11-19 7529500
Add minor release and add myself as an author.
