### 2019-11-19 a6ca35c
Make sure formatting is run.
