### 2019-11-20 ba33b85
Unload plugin if pytest is too old
