### 2019-11-20 75201db
Delete unused attribute
