### 2019-11-20 098415e
Bump hypothesis-python version to 4.44.3 and update changelog
[skip ci]
