+++ hypothesis-python/docs/changes.rst
@@ -21,6 +21,19 @@ Hypothesis APIs come in three flavours:
 You should generally assume that an API is internal unless you have specific
 information to the contrary.
 
+.. _v4.44.3:
+
+-------------------
+4.44.3 - 2019-11-20
+-------------------
+
+This patch avoids importing test runners such as :pypi`pytest`, :pypi`unittest2`,
+or :pypi`nose` solely to access their special "skip test" exception types -
+if the module is not in ``sys.modules``, the exception can't be raised anyway.
+
+This fixes a problem where importing an otherwise unused module could cause
+spurious errors due to import-time side effects (and possibly ``-Werror``).
+
 .. _v4.44.2:
 
 -------------------
