### 2019-11-20 74d49d5
fixing proceedings titles
