+++ paper.bib
@@ -2,9 +2,8 @@
   author    = {Koen Claessen and
                John Hughes},
   title     = {QuickCheck: a lightweight tool for random testing of Haskell programs},
-  booktitle = {Proceedings of the Fifth {ACM} {SIGPLAN} International Conference
-               on Functional Programming {(ICFP} '00), Montreal, Canada, September
-               18-21, 2000.},
+  booktitle = {{Proceedings of the Fifth ACM SIGPLAN International Conference
+               on Functional Programming (ICFP '00)}},
   pages     = {268--279},
   year      = {2000},
   crossref  = {DBLP:conf/icfp/2000},
@@ -18,9 +17,8 @@
 @proceedings{DBLP:conf/icfp/2000,
   editor    = {Martin Odersky and
                Philip Wadler},
-  title     = {Proceedings of the Fifth {ACM} {SIGPLAN} International Conference
-               on Functional Programming {(ICFP} '00), Montreal, Canada, September
-               18-21, 2000},
+  title     = {{Proceedings of the Fifth ACM SIGPLAN International Conference
+               on Functional Programming (ICFP '00)}},
   publisher = {{ACM}},
   year      = {2000},
   isbn      = {1-58113-202-6},
@@ -35,8 +33,7 @@
                Joakim Johansson and
                Ulf T. Wiger},
   title     = {Testing telecoms software with quviq QuickCheck},
-  booktitle = {Proceedings of the 2006 {ACM} {SIGPLAN} Workshop on Erlang, Portland,
-               Oregon, USA, September 16, 2006},
+  booktitle = {{Proceedings of the 2006 ACM SIGPLAN Workshop on Erlang}},
   pages     = {2--10},
   year      = {2006},
   crossref  = {DBLP:conf/erlang/2006},
@@ -50,8 +47,7 @@
 @proceedings{DBLP:conf/erlang/2006,
   editor    = {Marc Feeley and
                Philip W. Trinder},
-  title     = {Proceedings of the 2006 {ACM} {SIGPLAN} Workshop on Erlang, Portland,
-               Oregon, USA, September 16, 2006},
+  title     = {{Proceedings of the 2006 ACM SIGPLAN Workshop on Erlang}},
   publisher = {{ACM}},
   year      = {2006},
   isbn      = {1-59593-490-1},
@@ -91,8 +87,8 @@
                Chucky Ellison and
                Xuejun Yang},
   title     = {Test-case reduction for {C} compiler bugs},
-  booktitle = {{ACM} {SIGPLAN} Conference on Programming Language Design and Implementation,
-               {PLDI} '12, Beijing, China - June 11 - 16, 2012},
+  booktitle = {{ACM SIGPLAN Conference on Programming Language Design and Implementation,
+               (PLDI '12)}},
   pages     = {335--346},
   year      = {2012},
   crossref  = {DBLP:conf/pldi/2012},
@@ -107,8 +103,8 @@
   editor    = {Jan Vitek and
                Haibo Lin and
                Frank Tip},
-  title     = {{ACM} {SIGPLAN} Conference on Programming Language Design and Implementation,
-               {PLDI} '12, Beijing, China - June 11 - 16, 2012},
+  title     = {{ACM SIGPLAN Conference on Programming Language Design and Implementation,
+               (PLDI '12)}},
   publisher = {{ACM}},
   year      = {2012},
   url       = {http://dl.acm.org/citation.cfm?id=2254064},
@@ -174,9 +170,8 @@ Bdsk-Url-1 = {https://doi.org/10.3847/1538-3881/aabc4f}}
   author    = {Andreas L{\"{o}}scher and
                Konstantinos Sagonas},
   title     = {Targeted property-based testing},
[…]
