### 2019-11-20 64cff4e
Remove use of the word 'generator'
