+++ hypothesis-python/src/hypothesis/core.py
@@ -227,7 +227,7 @@ class WithRunner(MappedSearchStrategy):
         return "WithRunner(%r, runner=%r)" % (self.mapped_strategy, self.runner)
 
 
-def is_invalid_test(name, original_argspec, generator_arguments, generator_kwargs):
+def is_invalid_test(name, original_argspec, given_arguments, given_kwargs):
     """Check the arguments to ``@given`` for basic usage constraints.
 
     Most errors are not raised immediately; instead we return a dummy test
@@ -243,10 +243,10 @@ def is_invalid_test(name, original_argspec, generator_arguments, generator_kwarg
         wrapped_test.is_hypothesis_test = True
         return wrapped_test
 
-    if not (generator_arguments or generator_kwargs):
+    if not (given_arguments or given_kwargs):
         return invalid("given must be called with at least one argument")
 
-    if generator_arguments and any(
+    if given_arguments and any(
         [original_argspec.varargs, original_argspec.varkw, original_argspec.kwonlyargs]
     ):
         return invalid(
@@ -254,32 +254,32 @@ def is_invalid_test(name, original_argspec, generator_arguments, generator_kwarg
             "varkeywords, or keyword-only arguments"
         )
 
-    if len(generator_arguments) > len(original_argspec.args):
-        args = tuple(generator_arguments)
+    if len(given_arguments) > len(original_argspec.args):
+        args = tuple(given_arguments)
         return invalid(
             "Too many positional arguments for %s() were passed to @given "
             "- expected at most %d arguments, but got %d %r"
             % (name, len(original_argspec.args), len(args), args)
         )
 
-    if infer in generator_arguments:
+    if infer in given_arguments:
         return invalid(
             "infer was passed as a positional argument to @given, "
             "but may only be passed as a keyword argument"
         )
 
-    if generator_arguments and generator_kwargs:
+    if given_arguments and given_kwargs:
         return invalid("cannot mix positional and keyword arguments to @given")
     extra_kwargs = [
         k
-        for k in generator_kwargs
+        for k in given_kwargs
         if k not in original_argspec.args + original_argspec.kwonlyargs
     ]
     if extra_kwargs and not original_argspec.varkw:
         arg = extra_kwargs[0]
         return invalid(
             "%s() got an unexpected keyword argument %r, from `%s=%r` in @given"
-            % (name, arg, arg, generator_kwargs[arg])
+            % (name, arg, arg, given_kwargs[arg])
         )
     for a in original_argspec.args:
         if isinstance(a, list):  # pragma: no cover
@@ -289,9 +289,7 @@ def is_invalid_test(name, original_argspec, generator_arguments, generator_kwarg
             )
     if original_argspec.defaults or original_argspec.kwonlydefaults:
         return invalid("Cannot apply @given to a function with defaults.")
-    missing = [
-        repr(kw) for kw in original_argspec.kwonlyargs if kw not in generator_kwargs
-    ]
+    missing = [repr(kw) for kw in original_argspec.kwonlyargs if kw not in given_kwargs]
     if missing:
         return invalid(
             "Missing required kwarg{}: {}".format(
@@ -358,7 +356,7 @@ def get_random_for_wrapped_test(test, wrapped_test):
 
 
 def process_arguments_to_given(
-    wrapped_test, arguments, kwargs, generator_kwargs, argspec, test, settings,
+    wrapped_test, arguments, kwargs, given_kwargs, argspec, test, settings,
 ):
     selfy = None
     arguments, kwargs = convert_positional_arguments(wrapped_test, arguments, kwargs)
@@ -386,9 +384,7 @@ def process_arguments_to_given(
     search_strategy = TupleStrategy(
         (
[…]
