### 2019-11-20 6f7c736
Bump hypothesis-python version to 4.44.4 and update changelog
[skip ci]
