+++ hypothesis-python/RELEASE.rst
@@ -1,4 +1,0 @@
-RELEASE_TYPE: patch
-
-This patch adds some internal comments and clarifications to the Hypothesis
-implementation. There is no user-visible change.
