+++ hypothesis-python/docs/changes.rst
@@ -21,6 +21,15 @@ Hypothesis APIs come in three flavours:
 You should generally assume that an API is internal unless you have specific
 information to the contrary.
 
+.. _v4.44.4:
+
+-------------------
+4.44.4 - 2019-11-20
+-------------------
+
+This patch adds some internal comments and clarifications to the Hypothesis
+implementation. There is no user-visible change.
+
 .. _v4.44.3:
 
 -------------------
