### 2019-11-20 dd4ed91
Share more logic between explicit examples and standard given body
