### 2019-11-20 c7b931d
Test for deadlines in explicit examples
