+++ hypothesis-python/RELEASE.rst
@@ -1,0 +1,4 @@
+RELEASE_TYPE: minor
+
+This release commonizes some code between running explicit examples and normal test execution.
+The main user visible impact of this is that deadlines are now enforced when running explicit examples.
