### 2019-11-20 2420083
Remove unneeded arguments
