### 2019-11-20 f7d5dd9
Use a ConjectureData subclass so that we inherit all the right properties
