### 2019-11-20 bef75d9
No clear on py2
