### 2019-11-20 60ffbf2
Make sure formatting is run.
