### 2019-11-20 cc7ee7c
Remove helper function to validate kwargs, update release to include some markup and more explicit language. Remove separate helper function test file and include kwarg validation of rule() and initialize() testing inside test_stateful.
