### 2019-11-20 #1891 @danielskatz
👋  @DRMacIver & @Zac-HD  - at this point, can you archive the software somewhere (e.g., Zenodo) and let me know the DOI.  Please be sure the metadata (title and authors) associated with the archive match the JOSS paper.

Also, if the version number has changed, please let me know the updated version associated with the archived software.
