### 2019-11-20 #1891 @Zac-HD
We've been archiving at https://doi.org/10.5281/zenodo.1412597 for a while now!  That's via the github integration though, so it automatically picks up all contributors (same two main authors though).

The version number has been bumped, but we're actually mid-automatic-release... give it half an hour and we'll be at 4.45.0 😁
