### 2019-11-20 76fb06c
Bump hypothesis-python version to 4.44.5 and update changelog
[skip ci]
