### 2019-11-20 97a6af7
Bump hypothesis-python version to 4.45.0 and update changelog
[skip ci]
