+++ hypothesis-python/RELEASE.rst
@@ -1,4 +1,0 @@
-RELEASE_TYPE: minor
-
-This release supports :obj:`python:typing.Final` and :obj:`python:typing.TypedDict`
-in :func:`~hypothesis.strategies.from_type`.
