### 2019-11-20 dbbe001
Simplify validation tests
