+++ hypothesis-python/tests/cover/test_stateful.py
@@ -720,13 +720,9 @@ def test_invariant_precondition():
 
 def test_invalid_rule_argument():
     """Rule kwargs that are not a Strategy are expected to raise an InvalidArgument error."""
-
     with pytest.raises(InvalidArgument):
 
         class InvalidRuleMachine(RuleBasedStateMachine):
-            def __init__(self):
-                super(InvalidRuleMachine, self).__init__()
-
             @rule(strategy=object())
             def do_stuff(self):
                 pass
@@ -734,19 +730,11 @@ def test_invalid_rule_argument():
 
 def test_invalid_initialize_argument():
     """Initialize kwargs that are not a Strategy are expected to raise an InvalidArgument error."""
-
     with pytest.raises(InvalidArgument):
 
         class InvalidInitialize(RuleBasedStateMachine):
-            def __init__(self):
-                super(InvalidInitialize, self).__init__()
-
             @initialize(strategy=object())
             def initialize(self):
-                pass
-
-            @rule()
-            def do_stuff(self):
                 pass
 
 
