### 2019-11-20 a0f617d
Bump hypothesis-python version to 4.45.1 and update changelog
[skip ci]
