+++ hypothesis-python/RELEASE.rst
@@ -1,7 +1,0 @@
-RELEASE_TYPE: patch
-
-This patch improves the error message when invalid arguments are passed
-to :func:`~hypothesis.stateful.rule` or :func:`~hypothesis.stateful.invariant`
-(:issue:`2149`).
-
-Thanks to Benjamin Palmer for this bugfix!
