### 2019-11-21 #1891 @DRMacIver
> Also, if the version number has changed, please let me know the updated version associated with the archived software.

Ah this is something I meant to mention. Hypothesis versioning is very "If you don't like this version number, wait half an hour and there will be a new one" because of our automated release system. We've got a strong backwards compatibility policy and not much in this paper depends on the specific version, but I wonder if there's some way we could indicate that the paper is true of a version range? e.g. `>= 4.45.0, < 5.0.0`. It doesn't matter much and we can pick whichever version is out at point of paper release though.
