### 2019-11-21 0f51669
Add tests for some edge cases in health checks
