+++ hypothesis-python/RELEASE.rst
@@ -1,0 +1,4 @@
+RELEASE_TYPE: patch
+
+This release fixes a bug introduced in :ref:`Hypothesis 4.42.0 <v4.42.0>` which would cause occasional 
+:obj:`~hypothesis.HealthCheck.too_slow` failures on some tests.
