### 2019-11-21 #1891 @danielskatz
We need the DOI of a single version of the code, and the corresponding version number.

I suggest we use http://doi.org/10.5281/zenodo.3548997 and version 4.45.0

Ok?
