### 2019-11-21 #1891 @danielskatz
👋 @arfon - in the generated xml file, I see
```<unstructured_citation>In Praise of Property-Bsed Testing, ://increment.com/testing/in-praise-of-property-based-testing/, 2019, MacIver, David R.</unstructured_citation>```  I realize there's an ```a``` missing in the title, but the URL starting with ```://``` also seems odd - is this what is supposed to be in the XML?  (the URL in the pdf is fine)
