+++ paper.bib
@@ -57,7 +57,7 @@
 }
 
 @misc{PraiseOfPBT,
-  title = {In Praise of Property-Bsed Testing},
+  title = {In Praise of Property-Based Testing},
   howpublished = {\url{https://increment.com/testing/in-praise-of-property-based-testing/}},
   year={2019},
   author={David R. MacIver},
