### 2019-11-21 6a1131f
Speed up health check tests
