### 2019-11-21 #1891 @arfon
@danielskatz - this looks like a bug. Let's use Whedon to deposit metadata with Crossref here and I'll file a bug to investigate later.
