### 2019-11-21 #1891 @danielskatz
👋 @arfon - can you please check/fix the crossref xml data related to the url of the citation we discussed above?
