### 2019-11-21 #1891 @danielskatz
👋 @arfon - I'm going to close this now, but feel free to reopen it if you need to re the crossref issue
