### 2019-11-22 #1891 @djmitche
This is but a tiny sliver of the email I get from GitHub every day :)

Thanks to all for your hard work on JOSS :)
